Move pagination error observability out of domain
Strip `tracing`, `prometheus`, and the `PAGINATION_ERRORS_TOTAL`
process-global counter out of `domain::pagination_errors` so the domain
layer only constructs and returns `Error` envelopes. The pure
constructors lose their unused `PaginationErrorSource` parameter and
are renamed to `invalid_cursor_error` / `unsupported_direction_error`.

`PaginationErrorSource`, `record_pagination_error`,
`register_pagination_error_metrics`, and the metrics registration test
move into a new `crate::observability::pagination_errors` module. The
module is a top-level sibling of `inbound`/`outbound` so both the HTTP
adapter and the persistence error mapper can record the same
observation without crossing the inbound/outbound boundary forbidden
by the architecture lint.

`backend::register_pagination_error_metrics` still exposes the same
public registration entry point; only its internal implementation now
delegates to the observability module.
